finished fine-tuning
guys, if your changes don't reflect on the ausbildung server as you edit your code, you need to delete your browser data for the day that you're working on, or last one hour or whatever, depending on how long you've been working on it.

I was wrong about the server side error. It's just a browser caching issue.